refactor(text): refactor Text component location
