+++ src/components/commons/CardTitle/index.js
@@ -3,7 +3,7 @@ import { React } from 'react';
 import styled, { css } from 'styled-components';
 import { PropTypes } from 'prop-types';
 import breakpointsMedia from '../../../theme/utils/breakpointsMedia';
-import Text, { TextStyleVariantsMap } from '../../foundation/layout/Text';
+import Text, { TextStyleVariantsMap } from '../../foundation/Text';
 import Link from '../Link';
 
 const CardTitleWrapper = styled.div`
