+++ src/components/commons/Footer/index.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { PropTypes } from 'prop-types';
 import styled from 'styled-components';
-import Text from '../../foundation/layout/Text';
+import Text from '../../foundation/Text';
 
 const FooterWrapper = styled.div`
   position: sticky;
